SDUI: Added 'Work in progress' feature
https://trello.com/c/98Xcj19W

Currently all changes made by a user on MDS UI are stored in browser
and changes are sent to server when a user click 'Save changes' button.

This commit changes this behavior. From this commit all changes made by
a user are sent to server automatically and the user only commits changes
by clicking 'Save changes' button.

Also, from now on MDS UI a user could see in which entities made changes
and the user will be able resume edits or discard.

Change-Id: Ia02688e210e66c5ea26791d05d3dbcbd0f592b17