Bug Fix: Unable to enroll in campaign from tasks
Trello link: https://trello.com/c/13qiI3g7/1153

There are two fields "startTime" and "referenceTime" which needs to be defined
when enrolling an user, which are not so obvious for an end user to define.
The "referenceTime" field which was introduced as part of the bug
https://trello.com/c/ePvF6uhx/344, is used for campaigns when offset
is less than a day, which could be achieved with "starTime" field.
Hence, removing "referenceTime" field.

Added Javadocs.

Change-Id: I71d22ef00be31a0031c6c194b3b501e9c608f9db