BUG: Javascript Error in Tasks
https://trello.com/c/M1cOBrE4

The issue was that there was a coincidence of names (tab-pane - css
class name and angular directive at once). Removed file where the
directive was declared - angular-bootstrap.js. It seems that it is
not required for motech to run properly. I've found a few usages of
other directives included it this file (dropdownToggle and popover),
but I've haven't noticed that removing them broke down anything.

Change-Id: Ib278196f206269b263e70e711c8b65ec2dbfeefc