Adding fields to generated entity classes.
https://trello.com/c/JcNP0RMb

We use Javassist to add fields to the new class definitions. We also
add appropriate initializers based on the default value. This commit also
restructures the builder classes and adds methods which makes us ready for
bundle generations. Getters and setters are also generate for these fields.
MDSConstructorIT contains an integration test for persisiting these types.

Currently, the fields are not rebuilt with each commit to the metadata,
because we cannnot change registered metadata. This should get resolved
with transitioning entities to a separate bundle.

Change-Id: I9b07ae42c2dff01a7632f7515a1a901c19659827