MOTECH-1001 | Rationalize Scheduler API
 This commit removes the export of all scheduler packages.  To do that
the service classes where moved into a package called 'service' with the
impl in a subpackage.  The domain package was split up into two packages so
the classes used in the service api could be exported.  They were placed
into a new package called 'contract'.  The Private-Package config for
the bundle plugin was updated to include the remaining packages from
the module.

As part of the package changes all callers of these apis were updated.

Additionally to get the IT tests to pass a larger timeout was added.

Change-Id: Ie7255f0520b0b4663b6ba2ef75e22fd532244020